+++ src/RavenNest.Models/ClanSkill.cs
@@ -7,7 +7,7 @@ namespace RavenNest.Models
         public Guid Id { get; set; }
         public string Name { get; set; }
         public int Level { get; set; }
-        public decimal Experience { get; set; }
+        public double Experience { get; set; }
         public int MaxLevel { get; set; }
     }
 }
